scripts/install_dependencies.py: Update for GPL2017 and RTS build
